refactor(dialogue_manager): extract lifecycle.rs
Step 6/9. Move session-lifecycle methods + small public accessors
into dialogue_manager/lifecycle.rs:
- new (constructor)
- open_session, resume_session, close_session
- last_intent / last_assessment / classifier_identifier /
  last_extraction / extractor_identifier / last_comprehension /
  comprehension_identifier (used by --verbose)
- should_suggest_break (elapsed-time check)

These are all the methods on DialogueManager except respond_to and
respond_to_streaming, which stay in mod.rs for commits 7-8 (the
decomposition + turn.rs extraction).

Trims unused PedagogicalIntent / prompt_pack::self imports from
mod.rs's top-level use block (lifecycle.rs and apply.rs cover those
references now).

mod.rs: 2622 → 2330 lines.
lifecycle.rs: new, 245 lines.
